Added method
   public static getHome() 

in class IWeb2Config. If the "os.name" sytem proeperty value starts with "Linux" then the home value is determined by the value of the property "iweb2.home", which can be passed from the script. Otherwise, the returned value corresponds to the Windows convention of the book, i.e. "c:\iWeb2"